Further cleanup of formatting for `Home.js`
Again, this is pushing it closer to Ignia's current standards by introducing headers to break up sections, and aligning assignments. This isn't normally done for inline scripts as it requires a lot of unnecessary space which bloats the file size. But since these scripts now go through Uglify, that's not a problem.