Override HTML attribute type view in the Editor
For the Courseware, we'll want to expose some of the custom styles—such as for _Learn More_ and _Tips_—via the OnTopic Editor's WYSIWYG editor, which uses the CkEditor JavaScript library. To do this, we can simply override the `HtmlViewComponent`'s view, and it will automatically take precedence over the OnTopic version of the view.

This is just a barebones implementation to get started. Later, we'll be customizing this.