Introduced custom configuration for CKEditor
Since we're overwriting the CkEditor implementation for GoldSim, we need to provide a custom configuration script. This allows us to customize the toolbar, the styles, and the stylesheet.

We'll be placing this within `/Scripts/Views` since that's already configured to output individual scripts. I've extended the `gulpFile.js`, however, to allow scripts in subfolders to be included in that. That way, we can keep `CustomConfig.js` in its own `CkEditor` folder.