Introduced custom styles for CKEditor
Since the OnTopic Editor isn't part of the GoldSim website, and doesn't include its styles, it won't know how to render styles to look like GoldSim. Exposing the `contentsCss` allows us to configure that.

For now, these styles are the same as we use in OnTopic. Ideally, in a future commit, this will simply import the styles from e.g. `_typography.scss` and, possibly, `_contentBoxes.scss`, such that the editor uses the exact same styles as the website. That may require some refactoring to avoid off-target effects, though.