Introduced custom style set for CKEditor
While the previous commit (4cfe049) introduced a custom _stylesheet_ for rending content within the editor, this commit introduces a custom _style set_, for definining which styles should be available in the CKEditor _toolbar_.

For now, this is a barebones set. This will be updated with custom styles in a future commit.