Established `.editor` class for styles displayed within OnTopic
By configuring the `.editor` class, we can modify selectors that are intended to be within certain elements of the main site `_Layout.cshtml` to still display correctly within the editor, even though they'll be within a different context.

So, for instance, this includes the main body links, which are targeted for links within the `main` element. Since that won't exist within the OnTopic Editor, an alternate `.editor` selector is added to ensure they are still applied for the CKEditor implementation.