Added GoldSim specific styles to CkEditor
In a previous commit (ccb462d), we had established a barebones `StyleSet.js` for configuring the OnTopic Editor with a set of custom styles. We hadn't, however, added any custom styles to it. This update resolves that, by adding common styles used across GoldSim content pages. Not all styles are included, but the ones used most frequently are.