Introduce breadcrumbs to Lesson header
To aid in the navigation of the courseware—and to maintain consistency with the legacy Google Courseware implementation—added breadcrumbs to the Lesson header.

This necessitated following the `Unit` property on the `LessonTopicViewModel`. Otherwise, due to how mapped view models are cached by the `TopicMappingService` this would result in the `Parent` property's parents nto being mapped.

(This is due to a confusing side effect of how the internal caching works, and which normally doesn't occur except in cases where the same topic is mapped multiple times with different `[Follow()]` settings.)