Removed override of `@PageBodySection`
The Lesson template was overriding the `@PageBodySection`—but then defining `<div class="grid-x">`. Since the primary purpose of overriding `@PageBodySection` instead of using `@RenderBody()` is to get rid of (or customize) the `grid-x` declaration, this doesn't make much sense. Simply defining the body inline outputs the same markup and requires (slightly) less indentation and code.