Centralized `Layout` for courseware
Added a new `_Layout.cshtml` for the GoldSim Courseware. Notably, this allows us to override the navigation bar as well as the calls to action, without needing to do this on each page. This is important for the navigation bar since, unfortunately, it's not especially easy to override, and—unfortunately!—requires duplicating the header content. That's tolerable for this one `_Layout.cshtml`, but certainly not something we want to do for each course template.