Established more uniform breadcrumbs
There is now a basic breadcrumbs setup on the `_Layout.cshtml` which provides a link back to `Courses`. In addition, the breadcrumbs on the `Lesson.cshtml` don't repeat the current Unit, since that's being displayed as the title.

I don't love how this works, and may refactor the implementation to be more centralized later. For now, however, it at least unifies the interface.