Centralized header functionality in `_Layout.cshtml`
With the introduction of a central `_Layout.cshtml` for the courseware (d1ac20c), as well as a default breadcrumb (60e7026), there was a lot of duplication of markup and code between `_Layout.cshtml` and `Lesson.cshtml`, which overrides the `HeaderSection` in order to write the unit information.

It's not especially pretty to define a bunch of variables at the top of the `_Layout_ in order to conditionally implement this. Still, it's not too ugly and it helps centralize the complexity in the `_Layout`, thus keeping the individual templates simple.