Standardized Lesson navigation link treatment
Normalized treatment of `a` and `a:visited` since these should not vary for navigation. Used weight instead of underline for active state, similar to the original Google Courseware implementation. Used underline for highlight state, similar to e.g. the footer. Reduced the font size a bit since GoldSim prefers to use the full length Lesson names in the navigation; that will help reduce unnecessary wrapping (though, with many titles, that will be inevitable regardless).