Aligned Lesson navigation labels when wrapping
Ideally, the Lesson navigation would not wrap. GoldSim prefers that the full-length titles be used, however, in the navigation. If that's going to happen, then we at least need to make sure that the text aligns with itself. Otherwise, by default, the text wraps below the progress tracking icon, which looks unprofessional.

I'm not entirely clear on the logic behind this combination of `padding-left`, `text-indent`, and `margin-left` here. One might reasonably expect these should each be the same absolute value in order to align the text. But, in practice, that didn't work. This outcome is the result of trial and error. It seems to work, but I'd like to better understand the arithmetic to understand why.

I'd imagine there's e.g. a `rem-calc(4)` equivalent offset happening somewhere—perhaps on the icon margins—but I wasn't able to isolate anything via the debugger. As such, this necessitates further investigation and then hopefully, some documentation.