Reduced padding, font size for courses
As the courses are incredibly content heavy, and to be read more like an e-book than a marketing site, the normal font size and margins, while attractive, impacted the reading experience. This is especially true when also factoring in the left-hand navigation, which compresses the main reading panel. To account for that, I've reduced the font size and panel padding by a bit—though not so much, hopefully, to look inconsistent with the rest of the site.

We'll still want to review this internally from a design perspective, though it should be among the lesser of the compromises here.