Reduced font size to match Google Courseware
In a previous commit, we reduced the font size a bit to help account for the content density of the course content (64327c6). It was still a bit large, however, and especially for the navigation. With this reduction to 16px, we are now able to fit the same number of characters in a panel as the main marketing site, _despite_ now carving out room for the left-hand navigation. It also continues to look pretty consistent with the main site, and not nearly as cluttered as the Google Courseware site.