Restricted wrapping alignment to the navigation
In a previous commit, I ensured that when a list with progress tracking wrapped, it maintained alignment with the text, instead of wrapping around the progress tracking icon (c9addb0). This worked nicely for the Lesson navigation. Since some of the styles were specific to the Lesson navigation, however, but others weren't, this introduced an awkward gap for the Unit navigation.

This update restricts the previously shared portion exclusively to the Lesson navigation, so there aren't any off-target effects.