Redistributed columns on Lesson page
In a previous commit (747f35c) we established a 3:9 ratio for the navigation:body content. That produced a wider navigation than I preferred, but at the original font size there wasn't too much choice. With the smaller font size (64327c6, f298307), however, we can go back to a 2:10 ratio for both `medium` and `large` screen sizes. This maximizes the space available for the content.

There is still wrapping for the navigation titles—and especially as GoldSim prefers to use full-length navigation titles here. That wrapping is unfortunate, but within an acceptable range.