Accounted for tablets in Lesson navigation
While tablets aren't the target device for the courseware, we want the site to at least look acceptable on tablet. To that end, I've removed the lesson numbering and used that space to introduce a small left-hand margin for tablets. That allows us to maintain the 2:10 ratio while accepting smaller and smaller space for the left-hand navigation.