Introduced `LessonPagingViewComponent`
The `LessonPagingViewComponent` provides next/back paging for the lessons. This includes the view component, view model, and view, as well as custom styles (e.g., for padding, alignment), activation in the _Composition Root_, and implementation on the Lesson page template.

This implementation improves upon the Google Courseware implementation it is based upon, in that it provides navigation to the next or previous unit, and will return the user back to the syllabus if they're on the first lesson of the first unit, or the last lesson of the last unit.