Introduced `_TrackEvents` partial view
The `_TrackEvents` partial view loops through a `List` of the newly introduced `TrackingeventViewModel`, and writes each object as a Google Analytics event tracking call (`ga('send', { hitType: 'event', … })`). This makes it easy for server-side controllers or view components to relay events that can be tracked via Google Analytics on the client side.

(Note: These events _could_ be tracked via the server side as well, but when we do that we potentially lose some contextual data.)