Integrated `_TrackEvents` into courseware view components
Setup the courseware view components to be able to track custom events via Google Analytics. This involves wrapping the `NavigationViewModel<>` in a new view model with a `List<TrackingEventViewModel>` property, then relaying that property value to the `_TrackEvents` partial method.

This commit doesn't actually create any events, it just sets up the courses to be able to track events. Actual events will be added in subsequent events.