Introduced `StartUnit`, `EndUnit` tracking
These events will fire either the first time the user visits a lesson in a unit, or the first time they visit the last lesson which they have not yet visited.

As part of this, I introduced a new `IsUnitComplete()` helper method for evaluating the current state of the `Status{Unit}` cookie.