Updated OnTopic Editor to 4.3.1
This minor patch includes a number of bug fixes which impact GoldSim. Most notably, it fixes a breaking issue which prevented topics with an otherwise empty `TopicListViewComponent` from being saved; this primarily impacted any topics of type `ContentList` due to the generally unused `Category` dropdown list, which optionally allows content items to be grouped into categories.

Of relevance to GoldSim, this also includes a bug fix regarding how inherited values were defined and displayed. This prevents a confusing user interface issue where a deleted attribute value continues to be displayed. In fact, it's being displayed as a placeholder, but this was confusing because a) it was an explicit value, not an inherited value, and b) the font styling for the placeholder was indistinguishable from the normal font styling.