Conditionally display "GoldSim Website" in footer
On external sites, such as the Google Courseware, there is a link in the footer back to the main (marketing) website. On the main `www` server, however, we also have subsites that don't share the main navigation and may not otherwise have an obvious way back to the main site—such as the GoldSim Courseware.

To mitigate that, we've extended the `FooterViewComponent`'s view model to include a new `IsMainSite` property which is set in `FooterViewComponent`, and use it to conditionally display a "GoldSim Website" link in the footer if the current topic doesn't derive from "Root:Web".

As a nice-to-have, this also checks the `NavigationRoot` attribute—an existing feature which allows the main menu to repurpose the navigation from another root namespace. As a result, pages such as the Forms—which live in `/Forms`, but use the `/Web` navigation root—will _not_ display the "GoldSim Website" link in the footer.