Moved view models into newly established `Forms` area
There are a _lot_ of moving parts to the forms implementation. Currently, these are partially cordoned off using subfolders, such as `/Views/Forms` and `/Models/Forms`. But, even better, they can be placed within their own `/Areas` folder, as we do with e.g. the GoldSim Courseware.

While we're at it, this also gives us an opportunity to fix a design flaw in my initial implementation in which the partial models (i.e., those used to help compose the actual binding models) were placed in the root, while the actual binding models were placed in a subdirectory. It may make sense to have a `/BindingModels` subdirectory, but the nested objects which map to partial views should be _underneath_ that, not _above_ that.

As part of this, I also updated the namespaces to reflect the new folder structure, which necessitated changing some files outside of `/Forms` which maintained references to the `/Forms` section.