Moved views into newly established `Forms` area
In the previous commit (8b0a874) we moved the form-related view models into a new `/Forms` area. In this commit, we move the actual views into this area.

This also necessitates updating the `_ViewImports.cshtml` to include the registration for custom GoldSim tag helpers, as required for the `MetadataLookupViewComponent` used for dropdown lists in the forms.