Established route for new `Forms` area
Area routes are defined differently than normal controller routes. Unfortunately, the OnTopic Library doesn't (yet) have an extension method to aid with this. As such, the route must be declared using the long-hand format.

In addition, because the Forms rely on action-based routing for views, we need to define a route which will ensure actions can be defined. This is necessary because otherwise the `{*path}` route will be used, which doesn't override the action.

(An alternative would be to manually define the `action` route variable in each action of the `FormsController`. This is less error prone for future extensibility, though.)