Refactored view models into `Models` and `Models.ContentTypes`
Previously, to help declutter the `Models` namespace, we had established a separate `ViewModels`  and `Forms` folder—with the forms mostly being binding models. With the forms now moved to their own area (8b0a874), this distinction isn't quite as clear—nor necessary.

Given this, the general view models have now been moved (back) into the `Models` root, and the `ViewModels` namespace has been renamed to `ContentTypes`, as the view models it contains all map to specific OnTopic content types. This is a much more intuitive organization.

While I was at it, I moved the `TrackedNavigationTopicViewModel` into the Courses area, since it is used exclusively by the Courses.

As the namespaces were updated to reflect this reorganization, this necessitated changing other files—such as the controllers and view components—to ensure they were using the updated namespaces.