Picked up missing namespace renames
As part of the previous commit (e5a0187) we moved the general view models into the `GoldSim.Web.Models` namespace. For some reason, however, the search and replace for `cshtml` files failed to pick up a few instances of this. Oops!

(Normally, I'd roll back the commit and apply these changes in order to maintain a clean git history. Unfortunately, get doesn't seem to handle renames of files very well when a commit is rolled back, however, so I'm instead leaving that commit alone, and submitting this as a patch.)