Fixed area route for forms
In a previous commit, I established a route for the new forms area (52760c8). Unfortunately, in doing so, I set the `rootTopic`. This is needed when working with a `*path` in order to correctly determine the `CurrentTopic`—but it actually _interferes_ with `CurrentTopic` when using action-based routes.

That may be something we should address in OnTopic at a future date, as it's not immediately obvious that this should be the behavior. Nevertheless, that's how it currently works, and so this needs to be removed to ensure that the `CurrentTopic` is set correctly.

More importantly, _another_ limitation of OnTopic's support for areas is that it looks for topics in `/{area}/{controller}/{action}/{path}`—e.g., `/Forms/Forms/Trial`. That's not what we want here, as it's obviously redundant. This is something that needs to be fixed in OnTopic. Until then, however, I've added a (hopefully temporary) hack which removes the `controller` route value. This works since that value has already been used to lookup the `FormsController`, and nothing else in the pipeline (currently) relies upon it. That said, it's certainly not a _pretty_ solution, and we'll want to remove it later.