Applied (opt-in) bug fix for wildcard routes
There's a bug in ASP.NET Core routing (dotnet/aspnetcore#16579)  which prevents a wildcard routes to work if the route with the wildcard occurs _before_ the a more specific route—as we're doing with our form routes.

Fortunately, Microsoft introduced a fix for this in ASP.NET 3.1.301. Unfortunately, the fix requires that web applications opt-in to the correct behavior, presumably to maintain backward compatibility. This update applies that opt-in to `Program`.