Improved administration route
The `area` doesn't need to be explicitly defined for the administration route; this can be safely removed, as it's automatically defined based on the `areaName`.

While I was at it, I defined `LicenseController` as the default. This ensures that `/Administration` will route to a page, whereas previously the controller name was required.

(It may be even nicer to provide an administrative landing page which lists available administrative tools, but that's not a priority right now.)