Moved export service into Administration area
The export service is not suitable for general usage; it is largely written with the license export service in mind. If we have the need to generalize in the future, it would be nice to do so. But absent that need, it is more intuitively placed as part of the administration area.