Fixed path names in view comments for forms area
Views often contain a comment with the location of the view in order to aid in debugging. This is especially useful for partial views, or views which may not be sourced from the most obvious location. While refactoring the forms into a new area (0590b34), these comments were not updated, and thus still referenced their original location. This update fixes that.