Moved function definitions inside anonymous function
Previously, the core functionality had been separated out into a new stand-alone functions such as `initVideo()`. That was an improvement. But, in doing so, these were moved outside of the anonymous function, thus cluttering the global namespace. That's not the end of the world here since it's only for one page, but it's also not a best practice.

This update maintains the logical separation of functions, but moves them inside the anonymous function, consistent with our other scripts.