Established cache control header for static files
When we migrated to ASP.NET Core, the cache control headers that were previously in the `web.config's `<clientCache />` configuration stopped being supported. While we migrated other properties—such as the `<mimeMap />` extensions—from the `web.config` to the `Startup`, however, we didn't migrate the `<clientCache />` settings.

This update remedies that by establishing a 30 day cache for static files. This applies primarily to images, stylesheets, and scripts.

(Most content images now live on `media.GoldSim.com` which has a `web.config` with the `<clientCache />` settings. But static files which are part of the overall user interface are hosted on `www` not `media`.)