Removed `ToLower()` from canonical URL
The `/Sitemap` outputs the normal casing as defined in OnTopic, but the canonical URL within the markup uses `ToLower()`. This causes a conflict with Google, which interprets these as case sensitive. As a result, Google chooses the canonical URL over the sitemap URL. This means that a lot of pages are marked as duplicates, and makes it harder for us to assess how much of the sitemap is being honored.

Unfortunately, this is going to take a period for Google to recrawl the site and update the paths. It may also temporarily interfere with the search index. In the long-term, however, it should improve results by minimizing the number of perceived duplicates.