Reduce the markup output of the menu
The menu is a repetitive element on every page, and contributes to the overall page load. At some point, we may want to evaluate including this as an embeddable format. In the meantime, however, we should at least reduce the size of the markup generated.

Some of the markup was significantly larger than needed in aid of a readable template—e.g., by placing attributes on their own line. This can be fixed, but at a cost to that readability. To strike a balance between these, the variables are now separated out so as to minimize the logic that's integrated with the code.

This reduces the generated markup by about 0.6k. Not dramatic. But that adds up across every page. It also makes the generated output more readable.