Removed Bing Site Verification tag
I _believe_ that GoldSim's Bing Webmaster tools are now connected to Google Search Console, thus eliminating the need to maintain that. Even if that's not the case, however, the preference is to manage this via DNS. Unfortunately, it's not possibly to change to DNS verification when the site's already verified, so worst case if this invalidates the site then we'll just set it up with DNS verification at that point.