Fixed class assignments in menu
In a previous update, I refactored the `MenuViewComponent` view in order to reduce the size of the markup while maintaining readability (7a551b6). Unfortunately, in doing so, I failed to account for how variables are encoded, thus introducing attributes like `class=""selected""`.

There are a couple of ways of fixing this, but I opted to reorganize this again into more intuitive/descriptive variables which output exclusively the class or attribute value.

The downside of this approach is, in some cases, we may end up passing down an empty `class` attribute, which I had originally hoped to avoid. That said, this code remains lighter than the original, and is easier to follow than the last commit.