Migrated Application Insights to local package
This allows us to manage the package alongside other dependencies, reduces the amount of repetitive code on each page, and places the dependency under GoldSim's caching policy, which is more aggressive than Microsoft's.

In the future, it will also allow us to package the code as part of a larger JavaScript package, should we choose. (Currently, we bundle our custom scripts, but not vendor scripts, since many of them are page/template specific.)