Preload fonts to improve performance on modern browsers
Font declarations are included as part of the SCSS tree, and output to `Styles.css`. Unfortunately, this means that the browser needs to first download the entire CSS to discover the font locations, after which point it can begin downloading those. Preferably, browsers would be able to discover those earlier and begin downloading them in parallel, if practical.

This is permitted by use of the new(er) `rel="preload"` link. Not all browsers support it. But the ones that do will prioritize downloading these resources so that, by the time the CSS references them, they're already downloading—if not already available.