Import Google Fonts directly from Google
For reasons that aren't clear, we had copied the styles from Google's stylesheet locally.

This may have been in an effort to improve performance by allowing the fonts to be loaded upon discovery, instead of first requiring that the Google stylesheet was loaded. If so, however, that is no longer needed due to the use of `rel="preload"` (ef27fa5). Alternatively, it may have been an effort to trim the list down to the scripts we need by excluding e.g. the Vietnamese Unicode range. That also isn't necessary, however, as the browser will only download those fonts if it discover text within the target `unicode-range`.

Regardless, this isn't needed now, and can safely be replaced with the call to Google Fonts' CSS, as done here.