Removed legacy support for tables with a fixed column
Previously, as a fix for really wide tables, we had implemented the ability to "fix" the first column, while allowing the subsequent columns to scroll, similar to a technique Ignia had created for Ford. In practice, however, we haven't actually used this, and so there's no value in keeping it around as part of the stylesheet or JavaScript—and especially since the actual jQuery function was still wired up, even if it wasn't doing anything.