Removed unused Headroom implementation
Headroom is a useful script for hiding the header navigation as one scrolls down a page—and then displaying it again as one scrolls back up the page, thus preventing the need to scroll all the way to the top to navigate away from the page.

This was a nice feature. Unfortunately, we encountered some compatibility issues with how it interoperated with our header, possibly due to interoperability with Foundation's menu—I don't recall for sure. As a result, it's been commented out.

There's no sense keeping around files or code we're not actively using; if we need to, we can reimplement this at a later date.