Inherited `_foundationSettings` from `_settings`
The original `_foundationsettings.scss` file was copied from the `/scss/settings/_settings.scss` file that ships with Foundation Sites and then modified. While the changes that are GoldSim specific were clearly flagged inline, this still introduces a maintenance headache, as it repeats a lot of settings, and introduces ambiguity when underlying values have changed (e.g., did we just forget to flag it? Or was it intentional?).

To remedy this, I've inherited from the original `_settings.scss` file and carefully deduplicated variables. For this commit, I've maintained variables that were removed, added to the base file, or whose values have changed since we forked the file. This helps provide documentation of what's changed, in case anything creeps in (e.g., a changed value which, once removed, impacts our design). In a future commit, however, I will be removing these redundancies, favoring inherited values.

As part of this, I removed the headers of now-empty sections. I also fixed some of the formatting to be more consistent. (For example, the alignment was off in the last few sections, and some of the nested elements weren't properly aligned.)

As part of this, I had to correct the configuration of the `gulp-sass` plugin to correctly reference the Foundation Sites path; otherwise, root-relevative references from the `_settings.scss` (to e.g. `util/util`) won't correctly work.