Removed import of default `settings`
According to foundation/foundation-sites#8761, we shouldn't need to import `_settings.scss`, as all settings should fallback to their default values. That's not only convenient, but also avoids an issue where local variables from `_settings` need to be reapplied in the override—as we've discovered after inheriting from `_settings` (79c66e8).

That said, there is (at least) one case where this doesn't work as advertised. If the `$global-flexbox` setting isn't explicitly defined in our override file, then the `clearfix` mixin will fail, as it calls `$global-flexbox` without declaring a dependency. As such, we've had to add that back in using its default value to ensure everything compiles.

There aren't any other similar dependencies in the Foundation stylesheets we're currently using. Of course, if we expose additional modules in the future, we'll need to be aware of this possibility, and be prepared to import defaults if necessary.