Removed variables that had been newly introduced
A number of variables had been introduced since we originally forked `_settings` several years ago. These had been reflected in the comments as a placeholder until we could evaluate these more carefully (79c66e8). As these don't appear to have impacted the rendering of the site, and we have no immediate need to alter their values, I'm removing them from the override. This commit, however, provides a basic reference of what's changed in the meanwhile, should that be relevant in the future.