Removed variables that are no longer used by Foundation
A number of variables had been removed since we originally forked `_settings` several years ago. These had been reflected in the comments as a placeholder until we could evaluate these more carefully (79c66e8). As I've confirmed that we're not using these, and Foundation no longer relies on them, I've removed them from our settings file. This commit provides a basic reference of what's changed in the meanwhile, should that be relevant in the future.