Documented `floated-layout` style
It wasn't immediately obvious what the `floated-layout` styles were for, and they're not used in any of our templates. The only page they're currently used on is the product landing page. Since removing this impacts that layout, we're keeping it. To help avoid confusion, however, I'd added an inline comment to help document what the purpose of it is for.