Moved Braintree specific styles to a new view stylesheet
The only place where the Braintree overrides are used is on the `Payments.cshtml` page. The vast majority of users will never touch this page. The styles aren't that large, but given that there isn't (as much of) a disadvantage to loading multiple smaller assets in HTTP/2, it makes sense to move these to a separate view stylesheet. This also establishes a placeholder should we want to extend the styles for that template in the future.