Move Page-Level Navigation into separate component
Much of the current content from `_base` is actually specific to the Page-Level Navigation. That's not entirely clear based on the selectors, the comments, or the context. Regardless, this doesn't make much sense in the `_base`, which should be reserved for highly generalized styles that don't belong anywhere else. Given that, I've moved these styles to a new `_pageLevelNavigation` component style, which is a much more logical and consistent location.