Move Footer into separate component
After moving the Page-Level Navigation out of `_base` (365df66), most of the remaining styles are specific to the Footer. This doesn't make much sense in the `_base`, which should be reserved for highly generalized styles that don't belong anywhere else. Given that, I've moved these styles to a new `_footer` component style, which is a much more logical and consistent location.