Implemented Foundation card layout for applications
Previously, the homepage's Applications area used the `card` style, but didn't implement the `card-section` or inline image, instead preferring to use a background image for the image. That resulted in a lot of awkward in-between states where the title overlapped with the image.

This is fixed by moving inline, and adding a `card-section` wrapper for the content. This works _much_ better.

Because the (current) images don't map to any available key on the topic view model, I also added a helper function, `GetImagekey()`, to the `ApplicationContainerTopicViewModel`. This wasn't necessary when the image was handled by CSS, as the stylesheet handled that mapping.

This also allowed significantly simplifying the CSS, since we no longer need to adjust the top margin for various screen sizes—something which had added complexity, despite not working very well.