Introduced `GetContainerKey()` helper method for application container
The `Home.cshtml` included a couple of instances of unwieldy string parsing logic in order to strip the `Systems` suffix off of the `ApplicationContainerTopicViewModel.Key`. To reduce the repetition, and also simplify the logic in the view, I moved this to a helper function, `GetContainerKey()`, on the `ApplicationContainerTopicViewModel`. Much cleaner!