Added controls for application image positioning
As a background image, we were able to fine-tune the control of how the image was scaled, positioned, and cropped. We lost this when I moved the image inline (fb70185). By introducing a `<picture />` wrapper, I'm able to reintroduce at least some of this control.

(Conveniently, there was already a `picture` selector for these application cards in the SCSS, likely due to @ktrunkey implementing it this way prior to migrating to a background image. As such, I was just able to extend this existing functionality.)

While I was at it, I removed the legacy background positioning, since it is no longer needed.

Note: I'm still not able to get the image to be centered; I'll need to revisit that later.