Moved application padding from links to `card-section`
With the `card-section` container in place, we can move the padding from the list links to the overall container, thus ensuring the title adheres to the same limits as the list. This helps reduce an awkward intermediary where the title stretches beyond the boundaries of the list and bumps up against the border of the card.

(That said, there remains at least one awkward transition point where the `<nobr />` content is too wide for the area, and thus pushes off to the right side. I may need to revisit that—though it's unclear if that maps to any real world devices.)