Replaced `simplePagination` with basic jQuery script
We were previously maintaining a dependency on Flavius Matis' Simple Paging plugin for jQuery, but only using a very small part of it exclusively on the search page. Further, because it has limited options for styling, it was cumbersome to merge with our existing button styling. Rather than fight with the CSS, I just replaced it with a simple jQuery script.