Refactored search script as jQuery plugin
There was no good reason to keep the search script inline, so I moved it to an external script, where it can benefit from e.g., uglification and caching. While I was at it, I decided to repackage it as a basic jQuery plugin—which, really, is something we should be doing with a lot of these JavaScript files.

As a jQuery plugin, it's definitely a proof-of-concept. A more robust version would create the paging elements instead of relying on finding them via `id`—though those `id`s can, at least, be configured via the options. I'm also not certain that I'm following best practices in terms of maintaining scope; I'm currently relying on `bind(this)` to keep the scope on the jQuery plugin so that properties and methods can be referenced.

Regardless, this is a lot cleaner than having this code inline, and is something we should evaluate using for other scripts in the future.