Ensured scripts are deferred
For some reason, the main script was marked as both `async` and `defer`. That doesn't make sense. I believe in this case `defer` will take precident since it's second, but that's ultimately an ambiguity that should never have been introduced in the first place. As we want this script to be deferred, I've removed `async`.