Reintroduced input scripts into main stylesheet
In a previous commit, I moved all of the form related styles to a view-specific stylesheet, `Forms.scss` (f886cc5). That is preferred since there are a lot of styles, but they're only used on the handful of form pages on the site. That said, there are a couple of places where we use `<input type="text" />` on the site—most notably on the search results page—where we want to maintain consistency with the `[type='text']` styles. As such, the most basic definition is being reintroduced into the main stylesheet for general compatibility.

Note that this still need to be repeated in the `Form.scss`, unfortunately, because it `@include`s `foundation-forms`, which includes `foundation-form-text`, thus overwriting these definitions. We could fix this with `! important`, but that's not preferred—and, regardless, this is a very small amount of repetition. Should it become larger, we might consider using a mixin or include instead.

For now, however, this allows us to style (most) inputs without having to import all of the form-related styles.